feat: allow to link a directory with no manifest (#3695)
close #3691

+++ packages/local-resolver/src/index.ts
@@ -1,3 +1,5 @@
+import { existsSync } from 'fs'
+import path from 'path'
 import PnpmError from '@pnpm/error'
 import gfs from '@pnpm/graceful-fs'
 import { readProjectManifestOnly } from '@pnpm/read-project-manifest'
@@ -50,14 +52,22 @@ export default async function resolveLocal (
   try {
     localDependencyManifest = await readProjectManifestOnly(spec.fetchSpec) as DependencyManifest
   } catch (internalErr) {
+    if (!existsSync(spec.fetchSpec)) {
+      throw new PnpmError('LINKED_PKG_DIR_NOT_FOUND',
+        `Could not install from "${spec.fetchSpec}" as it does not exist.`)
+    }
     switch (internalErr.code) {
     case 'ENOTDIR': {
       throw new PnpmError('NOT_PACKAGE_DIRECTORY',
         `Could not install from "${spec.fetchSpec}" as it is not a directory.`)
     }
+    case 'ERR_PNPM_NO_IMPORTER_MANIFEST_FOUND':
     case 'ENOENT': {
-      throw new PnpmError('DIRECTORY_HAS_NO_PACKAGE_JSON',
-        `Could not install from "${spec.fetchSpec}" as it does not contain a package.json file.`)
+      localDependencyManifest = {
+        name: path.basename(spec.fetchSpec),
+        version: '0.0.0',
+      }
+      break
     }
     default: {
       throw internalErr
